backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
pyodbc==5.0.1
aioodbc==0.5.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
starlette==0.35.1

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Utils
python-dotenv==1.0.0
